guide-play-main/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
appshwnd==0.11
asttokens==2.4.1
atomics==1.0.2
attrs==23.2.0
auto-py-to-exe==2.42.0
bettercam==1.0.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.1
comtypes==1.2.0
contourpy==1.2.0
ctypes-window-info==0.15
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
dearpygui==1.10.1
debugpy==1.8.0
decorator==5.1.1
dxcam==0.0.5
Eel==0.16.0
executing==2.0.1
fast-ctypes-screenshots==0.10
FastCopyFast==0.52
fastimgconcat==0.10
ffmpeg-screenshot-pipe==0.10
filelock==3.13.1
flatten-any-dict-iterable-or-whatsoever==0.33
flatten-everything==0.41
flexible-partial==0.12
fonttools==4.45.1
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
game-overlay-sdk==1.0.1
get-rectangle-infos==0.10
getfilenuitkapython==0.11
getmonitorresolution==0.13
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
hackyargparser==0.12
idna==3.6
imageio==2.33.1
imutils==0.5.4
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
kthread==0.2.3
kthread-sleep==0.11
lazy_loader==0.3
list-all-files-recursively==0.12
logcontrol==0.2.1
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mjpeg-streamer==2023.12.25.post0
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multidict==6.0.5
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
Nuitka==2.0.3
numpy==1.26.2
nutikacompile==0.13
olefile==0.47
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
ordered-set==4.1.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pefile==2023.2.7
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prettytable==3.9.0
prompt-toolkit==3.0.43
proxy-tools==0.1.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
PyGetWindow==0.0.9
pyglet==2.0.10
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyinstaller-versionfile==2.1.1
PyMsgBox==1.0.9
pynput==1.7.7
pyparsing==3.1.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
pystray==0.19.5
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-osc==1.8.3
pythonnet==3.0.3
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinstyles==1.4
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
screeninfo==0.8.1
shapely==2.0.3
shlexwhichplus==0.10
simpleaudio==1.0.4
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
spawny==2.1.4
stack-data==0.6.3
subprocess-alive==0.10
subprocesskiller==0.11
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.12
termcolor==2.4.0
textdistance==4.6.1
threadmanager==0.1.0
tifffile==2023.12.9
tkextrafont==0.6.3
torch==2.2.0
torchvision==0.17.0
tornado==6.4
touchtouch==0.11
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
UltraDict==0.0.6
urllib3==2.2.1
vdf==3.4
wcwidth==0.2.12
websockets==12.0
wexpect==4.0.0
whichcraft==0.6.1
win32-setctime==1.1.0
windows-capture==1.0.60
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
